kernel/thread.c: allow multiple waiters on thread_join()
If multiple threads block on thread_join() for the same
thread, they will race to free the thread structure.

This change makes thread_exit wake only one waiter from
the retcode wait queue.  That thread will check if there
are other blocked threads and wake the next thread after
it collects the retcode and exits.

This change would be improved through the use of actual
thread reference counting.  While that could be done in
thread_exit/join with minimal impact, it still leaves the
calling interface questionably safe. If 'n' threads believe
they can thread_join(thread_X) but one of them does it much,
much earlier than the others, it will result in a use-after-free.

As such, this change solves an explicit bug and adds a test, but it
does not resolve the actual lifetime of thread objects.  It is likely
that thread_ts should have naive refcounting and access to threads
rely on a getter.